Fix data files for eeqa and sequence labelling
Instead of storing the data files inside the model code folder, the
configurations now point to the data folder in the root of the
repository.

Before this change, I have to manually copy the files from that folder
(which is the destination for the preprocess scripts) to the files
inside the models, which also meant we had duplicated data files all
over the place.

I might need to rewrite some of the documentation to account for this.

I also renamed some obsolete nomenclatures (e.g. ACE -> Sequence
Labelling).