self_critique: Update metrics and run files
It's not necessary to have two run files; one with an option is enough
since they're not doing anything special.

I've removed the the non-joint metric from self-critique because it's
not going to be used there.

The reconstruct metric is based on the joint one, but without the
classification part and without "kind", so there's no macro averaging.