Refactor genqa_joint and reconstruct preprocessing cripts
They had a lot in common (reconstruct is basically genqa_joint's answer
as the context plus the context truncation), so I created a common file
and abstracted away as much as possible.

I'm not convinced that unifying convert_instance to take a callable is
the best idea, but it's fine for now.