Fix mkdir parents usage.
Turns out that when you use .mkdir on the path to a file, it doesn't
create the parent directory, but instead tries to create a directory
with the file name. When there's a file already, it errors.

This means that the previous path.parent.mkdir usage was sort of
correct. It only created the parent, but if it were nested deeper, it
wouldn't create all parents. parents=True will take care of that.

There were two cases where the path was to a folder, so those were
unchanged.