self_critique: Disable useLibraryCodeTypes because of transformers
huggingface/transformers removed their py.typed file because they gave
up on properly typing their library. Having Pyright try to infer the
types is a waste of time, and everything performs much better when it's
off.

The problem with this is that pyright depends on the type information to
provide completion. I'm thinking of using pylsp for these kinds of
projects.