minimal(entailment): Remove PATH manipulation
This was done in order to import the metrics module for seq2seq, but
it's not needed for entailment. The metrics are defined within the
run_entailment.py script, so I just copied them here.

See `calculate_metrics`: just a small function; not worth complicating
the structure just to import it.