minimal(entailment): Fix shuffle problem
It was happening because I was trying to use the original data list,
whose order remains unchanged throughout the entire program, to
calculate metrics with the output of the model. Because of the
shuffling, the model data changed orders every epoch, which meant the
data labels became misaligned with the model data labels, causing the
issues in the metrics.

The fix is to use the labels from the model data instead of the original
data list. It's not pretty, but it works.

The next step is to apply the same fix to the Seq2Seq.

Note that there's still a bug, caused by the same principle: the `infer`
method tries to create an output with the original data plus the
predictions. This is broken for the same reason as the metrics. For now,
this can be prevented by not shuffling the inference data, but it should
be fixed properly.