minimal(seq2seq): Fix the shuffle bug
The reason is the same as in entailment (see history). However, the
implementation is very different.

Entailment only needs the label to calculate metrics, and that's needed
for the loss anyway, so no extra data is needed.

In seq2seq, we also need the context, answers and question_type for
FGCRCls to calculate the metrics, so we need to pass those through the
Seq2SeqDataset into the data loader.

The Dataset creates dicts with the data, and the data loader then
converts that into a dict of lists of that data. This means there's some
massaging necessary to get the data into the right format.

This resulted into a bunch of `list[dict[str, Any]]` and `dict[str, list[Any]]`,
which I don't like, so I created the TypedDicts Seq2SeqDatasetEntry and
Seq2SeqDatasetSeries (respectively) to make the types more clear. I
don't like TypedDicts, but they couldn't be dataclasses because the data
loader said so.

Anyway, now the shuffle works, and the metrics are calculated correctly.
And the seq2seq file is 60 lines longer than the entailment file.

I wanted to avoid all of this for entailment, but if I really want the
entire data to be present in the output predict file, I'm going to have
to at least pass the id with the labels, and at that point I might as well
pass the entire data.

That's a job for another day.