self_critique: Use latest version of TRL, and remove pdm.lock
The problem with the lockfile is that it's platform-dependent,
which means the pdm.lock generated on a Mac will add the MPS version,
and another generated on Linux will get the CUDA version.

We want both, depending on the machine, so we can't use the lockfile.
I think there's some work on PDM to enable conditional dependencies,
so maybe this will work at some point. For now, we suffer the 2 minutes
it takes for PDM to solve the dependencies without a lockfile.

In terms of the actual versions, it's not a big deal since the
runtime versions are pinned anyway.