chatgpt(paraphrase): Finalise scripts
I also made run_all.fish a generic script where you can pass the
script you want to run as a parameter. I did this because
run_all_paraphrase.fish and run_all_structured.fish were identical
except for the script that they were calling.

The problem is that I don't know how to call the shebang except
for ./path, which doesn't work when the path is a variable. Everything
here is a fish script, so I used it explicitly.