chatgpt: Add paraphrased data for entailment and reconstruct
Reconstruct is an odd case because doing this forces the model
to be bad because it's really hard for it to use the same words
as the gold output from the structured input.

However, it might still be worth it because the outputs are more
varied now. I need to take a closer look into a large sample before
I make the call to use this instead of the previous data, however.