self_critique: Add portal dependency
Portal is a nice data visualising tool that can work inside VS Code
and can do things like showing a seq of maps as a table, or showing
maps as a tree.

I also added a snippet in my VS Code to set it up easily, but it's
basically this: https://github.com/djblue/portal#api.