error_analysis(extraction): Use simple-parsing and save metrics
I'm going to write a script to run the classifier with multiple hyperparams,
and it's easier when I can supply a default config and override with CLI args,
which simple-parsing enables easily.

I'm also saving the metrics to a file so I can use it to compile a final result
(probably a CSV file) with that.