self_critique(scripts): Refactor ensemble_to_result.fish
Now it uses stdin if no parameter is provided, or uses the parameter
as the file name if it is.

ChatGPT tought me this trick of using 'cat -' to read stdin, so I can
just define the input variable instead of doing tricks like reading (cat)
to a string.

I'm probably to take this with me moving forward because I think this
combination is good. I don't have to create a temp pipe with psub like
when the scripts expects a file name, and I can't forget to <file when
the scripts only reads from stdin.

Usability win, yay.