self_critique(eval): New script to calculate "proper EM"
Previous version calculated macro-averaged across cause and effect.
This versions provides just EM (use eval.py for the rest), and gives
both cause and effect separately, and value for both. That is, not
averaged, but rather when both cause and effect are EM at the same time.

I call this "proper EM" because it's probably the most intuitive one,
but it's not what I found in other evaluation procedures. That being
said, this is the only one that can be used to calculate agreement.