evaluation: Make main calc.py a single-file calculation
It's what I'm using it for 90% of the time anyway, and writing the same
loop for all metrics and rewriting the output was getting old. The old
version is available as calc_mult.py in case I need single metric/multiple
data at some point, although I think that would be better served by a loop too.