self_critique(extract): Use greedy decoding during train evaluation
Using model.generate was taking too long (it's easily 100x slower than
just model(...) + argmax), which meant training was taking 1.5x longer
than it needed to.

This isn't perfect because it means we're doing early stopping (based on
the F1) on the greedy decoding instead of whatever we set the generation
to be, but it will do. I don't expect the F1 to be too different between
different strategies anyway.

This will still influence the eval/prediction after training.