self_critique(extract): Add note on whether the git repo is dirty
"dirty" means that there is stuff that hasn't been committed. This
serves to note that when the script ran, there was uncommitted stuff, so
just reverting to that commit might not be enough to get the same
result.

Ideally I'd commit before running everything, but that isn't always
possible. Although I could come up with some thing for a "temporary
commit" that commits everything before running the script, and just
resets to the previous commit afterwards. That sounds a bit complex, but
it would be the ideal.