self_critique(extract): Finalise first step of generate rewrite
Now I need to test and optimise.

I also wanted to make generate return the logits, but that would be like
model.generate, where it would only return a list of logits, and it
would require some manipulation to get them to a format usable with
CrossEntropyLoss. Given how the forward pass is already reasonably fast,
I'll keep as it is.