evaluation(data): Add new "lite" files and a README
The README covers the different types of data files, including the new
lite files.

The lite files are the tagged files without the junk that isn't useful
anymore, like the separate clauses. These pollute the file a lot when
I'm trying to read it, so I finally go annoyed enough to do something
about it.fd