annotation: Rework how the Tag enum is used
I'm not saving the annotation ratios anymore, so there aren't any cases
of Tag being a key, only a value. Because of that, I can just pass
`default=str` to the `json.dumps` method and that's taken care of. The
default seems to be using the Enum.value, which is good.