annotation: Move Logout button to only the Admin page
It's bugged on Start Page. Logging out would clear the query param from
the internal Streamlit dictionary, but not from the URL, so a refresh
would auto-login the user.

The users don't actually have to log out, so logout was mostly useful
for testing anyway. So now it's on the Admin page, and if it bugs out, I
can manually edit the URL.