annotation: Automatically assign data files to new Prolific IDs
The idea is that the mapping file would start with all the keys mapping
to null, and as users come in, they get assigned the next free key.

I'm logging the shit out of this (might add more info later on) because
I'm afraid of bugs here.

Also, I'm currently raising an exception. I'll probably try to present a
nice error to the user later, but this is a fatal error anyway.