annotation: Make pyright strict
Some cases have to be manuall ignored:
- st.write is incomplete, somehow, but my usage is equivalent to using
st.markdown since I'm always writing strings, and st.markdown works
fine.

- The import from evaluation isn't going through. I think it should be
possible to fix that with py.typed, but for now this will do.

- Some of the JSON stuff gives me "unknown" types, which is
understandable, but it clashes with my attempt to manually validate
that. My solution for now is to ignore this.