annotation: Print error message when the user couldn't be allocated
The only problem with this is that some of these cases (like getting the
latest index) shouldn't fail because of this, as they are called only
when on the annotation page, which guarantees that the data file exists.

I couldn't think of a better way to do this, though, so I'm going to be
using these tautological checks.