evaluation(st): Improve documentation, showing the docstring
I just realised that I can have the first line of the docstring and the
rest as an epilogue. That helps _so much_ in cases like specifying
input/output formats.

The one downside is that I can't use the RawHelp formatter (which is
necessary for this, otherwise the text gets bunched up as a single line)
and the Default formatter at the same time. That's not a big deal, since
I can add %(default)s to the help strings to get automatic defaults, and
if it really is a problem, I can solve it by multi-inheritance. It's
ugly, but it'd work. I haven't seen a case that warrants it yet.