self_critique(scripts): Make Fincausal eval default script use JSON
Now I can directly take model outputs, or thruogh reshape_output.fish if
not exactly the same. Skips the dance of creating ref/pred CSV files and
so on.

As a courtesy from argparse.FileType, it can also take stdin directly,
so we can easily chain reshape_output.fish and this for seamless
evaluation if required.