sequence_labelling: Always flush tokens at the end of a block
So, storytime:

This tiny change (move stuff out of the if so it always runs) solved a
bug that took me an afternoon.

The problem here is that this works with FCR, both the old data and the
newly converted one from tagged. It wasn't working for FinCausal, and
the error is that somehow the sequence was huge. It said that the
maximum was 512, and that I was giving it 20k or something.

I eventually tracked down to this: the original idea was just keep every
single token together, and just chop between max_len segments. I'm not
sure why this was done, since there was no concept of when one example
ended and the next started, but I suppose that doesn't make a difference
when evaluating what is essentially a token-level classification task.

The problem is that this "len(list_tokens) + 2 <= max_len" condition
works if each instance is relatively short and you're never jumping too
much in the number of tokens. This was fine for FCR, but not for
FinCausal.

You see, FinCausal has some _long_ instances. Or at least, long after
tokenised, considering it has a ton of numbers and so on. This meant
that this condition was skipped, and the list just kept growing until
after the loop, when it was flushed all it once. Creating a tensor of
batch_size x way_too_many_tokens.

Flushing after every block fixes this. It means that we now have decent
sized tensors.

There is one problem left here: I don't like how it does block
"detection" here. Essentially, if it finds an empty line, it creates a
block by flushing the existing tokens. I would like to do a much simpler
"split by \n\n to get the blocks, which you then splitlines() to get
each line and process that". I think that's much better, but would
require some larger changes to this. I don't like the flush_tokens stuff
either.

But I'll minimise the changes here so I can move on to more important
things. I spent way too much time here already.

Anyway, this solves the bug, and hopefully lets me run this thing.