data(maven-ere): Add gzipped train and valid data files
MAVEN-ERE is similar to FinCausal is that the test set is blind, and we
can't access it directly, so we're ignoring it.

Also, MAVEN-ERE is of a decent size. The processed train set is almost
100M, so the conversion script allows for compression. This means I'll
have to change all the json loading functions to cope with it, but it
shouldn't be a big deal because Python's stdlib includes a gzip module,
and it's pretty easy to use.