self_critique: Change how the metric classes are used
Now that there is a bunch, I figured there would be a better way to
access them. This seems cleaner to me. If somehow a wrong name gets the
(there's a check in the seq2seq script too), it will just fail as a
KeyError, which is fine with me.

Unfortunately, I can't transform the classes into simple functions
because that would break too much code. I wish I could, but this
temporary instantiation into "private" method invocation will have to
do.