chatgpt(extraction): Add MAVEN straight mode
Cheating on the parsing a little bit. I think that if I repeat the same
answer for cause/effect, when the metrics end up macro-averaging, it
will be the same as cause-only approach.

I'm only doing this hack here because GPT extraction isn't our focus; we
just want some baseline numbers. The FLAN-T5 extraction/RL stuff uses a
specific metric.