self_critique(rl): Add configurations for MAVEN with F1 and SentenceTF
The thresholds likely need to be tuned somehow - we don't have human
data on MAVEN to callibrate - so this is probably really bad.

Also, MAVEN's F1 threshold needs to be halved since the F1 score for
effect is always going to be zero, making the max F1 0.5 instead of 1,
and I'm too lazy to fix this properly and adjusting the threshold is
easier.