data(maven): Use sliding window to generate multiple examples per entry
I was taking just the first causal sentence, but it's possible to
extract more by using a sliding window and being careful not to include
previous sentences into the context.

I also added some limits to the context size by adding a hard limit on
the distance to the start and end indices, in case the random generated
ones go too far. 2 seems like a decent number, but I still need to test
this further.

Finally, I'm a little paranoid that I'm missing something and can
generate multiple instances with the same context, which would be
_really bad_. So there's an assert at the end just for that.