data(maven): Refactor the indices logic a little further
I _think_ this is concise and still readable. I might change my mind and
go back on it.

It's the -1 index the bothers me the most because it's artificial (it's
-1 just so that I can +1 to become 0 without a condition), but
len(sentences) is also artificial and it didn't bother me.

I might also backtrack on using slicing for the previous/current/next
part instead of one declaration per line.