self_critique(config): Increase input len and decrease batch for MAVEN
MAVEN's contexts are long, and we were being too harsh on the input data
filtering to fit on shorter contexts. So 512 it is, with longer examples
(including the prompt) being removed from the data.

Because we double the length, we also need to decrease the batch size. 4
might be too conservative; we'll see.