agreement(scripts): Rename agreements metrics module to avoid clashes
There is a "metrics" module in chatgpt, and the LSP was clashing the
two. This isn't great, but I don't want to make agreement a package of
its own (although I probably should), so this will do for now.

"This will do for now" is like my catch phrase at this point.