annotation(scite): Add kappa calculation script
This one is simpler because I know Hanqi and I are annotating the same
samples, so I don't have to do anything clever...

But I did anyway, typing _everything_ I could. No dicts in sight after
parsing. I even tried to do a SortedList type so I could be sure the
inputs of calc_kappa were sorted before I checked their equality, but
I'll be satisfied with knowing I sort parse(), and that will have to be
enough, unless I figure out a way to make this concise in Python.