agreement(scripts): Have eval_f1 optionally receive a threshold
It can calculate the binary labels directly now, instead of relying on a
separate step. Also, the binary label should be _pred_, not the
continuous value.

The threshold would come from somewhere else. For example, it would be
obtained by runnning find_threshold.py on training data, and using the
found threshold for the dev evaluation, effectively "training" the F1
classifier.