data(evaluation): Add script to run element-wise metrics
Most of the metrics here (F1, BLEU, BLEURT, BertScore, ROUGE-L) are
usually run on a database level, but we wanted to see how well they
correlated with human annotations. So, we need to calculate them on an
element-wise basis.

F1 was already done (see agreement dir), but I wanted it to be here too,
as it was more convenient.

NB: This takes a while to run. On my machine, Tensorflow wouldn't find
the GPU, and running all the metrics on CPU took many minutes (idk,
10+).