self_critique(rl): Parametrise evaluation prompt for run_reward
Before it was hardcoded to the now "combined" variant. This change
brings run_reward.py to par with extract_train.py here and classifier.py
on evaluation.

NB: the prompt is tied to how the evaluator/classifier was trained, so
this setting must be set to the same as the model had during training.
However, there's currently no way to set this from a config file inside
the saved model, so this must be set in the CLI when running
run_reward.py. If this is set incorrectly, the model will behave badly.