self_critique(scripts): Replace run_reward fish script with python
With better CLI handling, and more flexibility.

For example, now there's the --cpu option that not only uses CPU instead
of CUDA, for those cases where CUDA is busy, but it by default sets a
very high (1e9) batch size so we can load the whole dataset to memory.
The batch_size is still customisable in case I'm somehwere with less
RAM.