self_critique(rl): Refactor Evaluator init out of extract_train's main
That huge block in main felt wrong. Also, this instance is surprising
cleaner with a chain of if's rather than a match block. This is because
of the reward_model condition, that it must be present for
SentenceTransformer and Reward but not for F1, which makes it awkward.

At least we can rely on the typer checker to check for exhaustiveness,
since the type of evaluator_type would be Never if all of the variants
were checked by the ifs. We can guarantee this by using assert_never,
which is new in Python 3.11, but we can use typing_extensions if we need
to go back to 3.10 for some reason. Anyway, assert_never tells the type
checker to ensure that that variable is of type never, and gives a type
checking error if one of the variants isn't checked. I bit contrived,
with match still being better, but in this case it's good enough.