Add indication of projects that need newer Pythons
The assumption here is that the "base" version is 3.10, which is
specified in all the other pyproject.toml files, with these two being
the exception.

Both instances come from the typing module (override for annotation,
assert_never for self_critique), so they can be easily supplied by
typing_extensions if I need to go back to 3.10.

However, I'm reasonably committed to 3.12, since everything seems to be
working well, so I'm not going to do that. If I need to drop down a
version or two in the future, likely because of a library, at least I
have those comments indicating what to do.