evaluation(scripts): Add new script to run iterative weak reward
The idea is this:
- We train an evaluator on a training set
- We apply this evaluator on the next data split, generate labels and
confidences
- Filter the dataset andd add to the existing training set
- Train a new evaluator with the new training set

And so on. Is it the best way to do this? Probably not. But it was quick
and easy to implement (the damn file splitting was the hardest part, and
I had to circle back on that once I found a bug). So it goes.